requirements.txt
# 批量打印工具依赖
# Windows 上需安装 Microsoft Word/Excel 以使用 COM 打印
pywin32>=306
# 用于解析文档类型、路径等
python-docx>=1.1.0
openpyxl>=3.1.0
# PDF 仅用于检测与可选校验（实际打印可用系统默认程序）
PyPDF2>=3.0.0
# Web 界面
flask>=3.0.0
# 在线签名：可选 MySQL 共享上传列表（配置 MYSQL_HOST 后启用）
PyMySQL>=1.1.0
python-dotenv>=1.0.0
# 无 7-Zip 命令行时解压 .7z（.rar 建议安装 7-Zip 或使用 UnRAR）
py7zr>=0.20.0
